Updating penman examples and code
