+++ Penman.Blazor.Quill.sln
@@ -17,7 +17,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Penman.Blazor.Quill", "src\Penman.Blazor.Quill\Penman.Blazor.Quill.csproj", "{709BDDB9-F517-4762-89AC-4CF6C1229BEA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorWasm", "samples\BlazorWasm\BlazorWasm.csproj", "{9D647BCF-28EC-428F-B6F2-01D7BE9673B0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorWasm", "samples\BlazorWasm\BlazorWasm.csproj", "{536D1DAF-91A2-489C-8540-0EA204AFEE64}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,17 +33,17 @@ Global
 		{709BDDB9-F517-4762-89AC-4CF6C1229BEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{709BDDB9-F517-4762-89AC-4CF6C1229BEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{709BDDB9-F517-4762-89AC-4CF6C1229BEA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9D647BCF-28EC-428F-B6F2-01D7BE9673B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9D647BCF-28EC-428F-B6F2-01D7BE9673B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9D647BCF-28EC-428F-B6F2-01D7BE9673B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9D647BCF-28EC-428F-B6F2-01D7BE9673B0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{536D1DAF-91A2-489C-8540-0EA204AFEE64}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{536D1DAF-91A2-489C-8540-0EA204AFEE64}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{536D1DAF-91A2-489C-8540-0EA204AFEE64}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{536D1DAF-91A2-489C-8540-0EA204AFEE64}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{AEA7320C-E669-4009-B877-20E49D1D1687} = {666C8D8F-0108-4D22-9FA8-E1C65E42B1E2}
-		{9D647BCF-28EC-428F-B6F2-01D7BE9673B0} = {666C8D8F-0108-4D22-9FA8-E1C65E42B1E2}
+		{536D1DAF-91A2-489C-8540-0EA204AFEE64} = {666C8D8F-0108-4D22-9FA8-E1C65E42B1E2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9F1B7D95-9415-4F7A-849C-FBF83101BECF}
