+++ samples/BlazorServerSide/Pages/_Host.cshtml
@@ -9,13 +9,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Penman.Blazor.Quill Server Example</title>
     <base href="~/" />
-    
-    <!-- Quill stylesheet -->
-    <link href="//cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet">
-    <link href="//cdn.quilljs.com/1.3.6/quill.bubble.css" rel="stylesheet">
 
     <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
     <link href="css/site.css" rel="stylesheet" />
+    
+      
+    <!-- Quill and Penman.Blazor.Quill stylesheet -->
+    <link href="//cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet">
+    <link href="//cdn.quilljs.com/1.3.6/quill.bubble.css" rel="stylesheet">
     <link rel="stylesheet" href="_content/Penman.Blazor.Quill/penman-blazor-quill.css" />
 </head>
 <body>
