+++ samples/BlazorWasm/wwwroot/index.html
@@ -6,15 +6,17 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
     <title>Penman.Blazor.Quill Wasm</title>
     <base href="/" />
-
-    <!-- Quill stylesheet -->
-    <link href="//cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet">
-    <link href="//cdn.quilljs.com/1.3.6/quill.bubble.css" rel="stylesheet">
-
-
     <link href="css/bootstrap/bootstrap.min.css" rel="stylesheet" />
     <link href="css/app.css" rel="stylesheet" />
+
+    <!-- Quill and Penman.Blazor.Quill stylesheet -->
+    <link href="//cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet">
+    <link href="//cdn.quilljs.com/1.3.6/quill.bubble.css" rel="stylesheet">
+    <link rel="stylesheet" href="_content/Penman.Blazor.Quill/penman-blazor-quill.css" />
+
     <link href="BlazorWasm.styles.css" rel="stylesheet" />
+
+
 </head>
 
 <body>
