+++ src/Penman.Blazor.Quill/wwwroot/penman-blazor-quill.css
@@ -4,6 +4,6 @@
 }
 
 .toolbar.sticky {
-    z-index: 2;
+    z-index: 2000;
     background-color: rgba(255,255,255,0.9);
 }
